3dgsVC/requirements.txt
# 3DGSMR依赖包
# pip install -r requirements.txt

# PyTorch (请根据CUDA版本选择)
torch>=2.0.0
torchvision>=0.15.0

# 数值计算
numpy>=1.21.0
scipy>=1.9.0

# 数据处理
h5py>=3.8.0

# 配置管理
pyyaml>=6.0

# 进度条
tqdm>=4.64.0

# 图像处理和评估
scikit-image>=0.19.0

# 可视化 (可选)
matplotlib>=3.5.0

# 类型检查 (开发用)
# mypy>=0.990
# typing-extensions>=4.0.0
